requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.24.0
boto3==1.35.71
botocore==1.35.71
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
django-storages==1.14.4
gunicorn==23.0.0
h11==0.14.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
mysqlclient==2.2.6
olefile==0.47
packaging==24.2
pillow==11.0.0
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
opencensus
azure-identity
django-environ